FILE: moves non-exported content to 'internal' folders
