+++ libs/spirit-islander/config/util/src/lib/internal/combo-analyzer.ts
