+++ libs/spirit-islander/config/util/src/lib/internal/combo.interface.ts
