+++ libs/spirit-islander/config/util/src/lib/internal/get-possible-combos.spec.ts
