+++ libs/spirit-islander/game-setup/util/src/lib/create-game-setup.ts
@@ -1,8 +1,8 @@
 import { Difficulty, Options } from '@atocha/spirit-islander/shared/util';
 import { Config, getValidCombos } from '@atocha/spirit-islander/config/util';
 import { GameSetup } from './game-setup.interface';
-import { selectRandom } from './select-random';
-import { selectBoards } from './select-boards';
+import { selectRandom } from './internal/select-random';
+import { selectBoards } from './internal/select-boards';
 
 export function createGameSetup(config: Config): GameSetup {
   const { players, expansions, spiritNames, boardNames } = config;
