+++ libs/spirit-islander/game-setup/util/src/lib/internal/select-boards.spec.ts
